refactor(qa): #368 contrast-check thin 분리 — lib/qa/ 구현 + routing wrapper
coding-standards thin command pattern 준용 (#368 범위 일부):
- lib/qa/contrast-check.ts (신규): 순수 함수 4종 + COLOR_PAIRS + runContrastCheck 구현.
- scripts/ci/contrast-check.ts (120→29줄): thin routing wrapper.
ci:contrast 동작 보존 (exit code / warnings). 회귀 직검: typecheck 0 / ci:contrast EXIT 0.

#368 잔여: .github/workflows/qa.yml gate-1~5 PR 통합 (medium 회귀, 별도 진행).

Refs #368

🗿 MoAI <email@mo.ai.kr>
